refactor: Enhance platformio configuration and introduce new drivers
- Consolidated common platform configurations in platformio.ini for STM32 and native environments, improving maintainability.
- Added new LSM6DSOW and MA330 drivers with comprehensive register mapping and functionality for sensor integration.
- Introduced MT6701 driver with support for both I2C and SPI communication, enhancing flexibility in sensor applications.
- Updated existing driver headers to improve organization and clarity, ensuring better integration with the overall architecture.
- Enhanced error handling and type definitions across new drivers for improved robustness and usability.

+++ lib/hortor/src/math/profile.h
@@ -10,7 +10,7 @@
 
 #include <Arduino.h>
 
-#include "math.h"
+#include "math/math.h"
 #include "noncopyable.h"
 
 namespace hortor::math {
@@ -66,7 +66,7 @@ class Profile : public hortor::Noncopyable {
   void Process(float dt);
 
   /// @brief 当前轨迹位置 [pulses]
-  int32_t position_trajectory() const { return static_cast<int32_t>(pos_traj_); }
+  float position_trajectory() const { return pos_traj_; }
 
   /// @brief 当前轨迹速度 [RPM]，用于写入寄存器
   float velocity_trajectory_rpm() const { return vel_traj_rpm_; }
